+++ packages/dso-toolkit/src/components/legend-item/legend-item.args.ts
@@ -10,15 +10,17 @@ export interface LegendItemArgs {
   disabledMessage: string;
   dsoMouseEnter: HandlerFunction;
   dsoMouseLeave: HandlerFunction;
-  dsoRemoveClick: HandlerFunction;
+  dsoActiveChange: HandlerFunction;
+  active: boolean;
   label: string;
-  removable: boolean;
 }
 
-export const legendItemArgs: Omit<LegendItemArgs, "dsoMouseEnter" | "dsoMouseLeave" | "dsoRemoveClick" | "label"> = {
+export const legendItemArgs: Omit<
+  LegendItemArgs,
+  "dsoMouseEnter" | "dsoMouseLeave" | "dsoActiveChange" | "label" | "active"
+> = {
   disabled: false,
   disabledMessage: "",
-  removable: false,
 };
 
 export const legendItemArgTypes: ArgTypes<LegendItemArgs> = {
@@ -40,16 +42,16 @@ export const legendItemArgTypes: ArgTypes<LegendItemArgs> = {
     ...noControl,
     action: "dsoMouseLeave",
   },
-  dsoRemoveClick: {
+  dsoActiveChange: {
     ...noControl,
-    action: "dsoRemoveClick",
+    action: "dsoActiveChange",
   },
   label: {
     control: {
       type: "text",
     },
   },
-  removable: {
+  active: {
     control: {
       type: "boolean",
     },